% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{GDCdisease}
\alias{GDCdisease}
\title{GDC projects list}
\format{A list with 39 elements}
\usage{
GDCdisease
}
\description{
GDC projects list
}
\examples{
\dontrun{
data("GDCdisease")
}
}
\keyword{internal}
