% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{data}
\name{gencode.v36.annotation.genes}
\alias{gencode.v36.annotation.genes}
\title{GENCODE v36 gene information}
\format{
A Granges object
}
\source{
Downloaded from GENCODE v36
https://www.gencodegenes.org/human/release_36.html
Comprehensive gene annotation and filtered to genes
}
\usage{
data("gencode.v36.annotation.genes")
}
\description{
GENCODE v36 gene information
}
\keyword{datasets}
