% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{data}
\name{gene.location.hg19}
\alias{gene.location.hg19}
\title{Biomart hg19 gene information}
\format{
A table
}
\source{
Downloaded with biomart
}
\usage{
data("gene.location.hg19")
}
\description{
Biomart hg19 gene information
}
\keyword{datasets}
