% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{data}
\name{gene.location.hg38}
\alias{gene.location.hg38}
\title{Biomart hg38 gene information}
\format{
A table
}
\source{
Downloaded with biomart
}
\usage{
data("gene.location.hg38")
}
\description{
Biomart hg38 gene information
}
\keyword{datasets}
