% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{data}
\name{glioma.gcimp.model}
\alias{glioma.gcimp.model}
\title{gcimp RF model}
\format{A random forest model with 276 samples and 145 predictors classifying into 3 classes}
\source{
RF model created from DNA methylation signatures retrieved 
from \url{www.cell.com/cell/abstract/S0092-8674(15)01692-X}
}
\usage{
glioma.gcimp.model
}
\description{
A RF model able to classify DNA methylation samples in to GCIMP groups
}
\examples{
data("glioma.gcimp.model")
}
\keyword{internal}
