% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{data}
\name{glioma.idh.model}
\alias{glioma.idh.model}
\title{IDH RF model}
\format{A random forest model with 880 samples and 1205 predictors classifying into 6 classes}
\source{
RF model created from DNA methylation signatures retrieved 
from \url{www.cell.com/cell/abstract/S0092-8674(15)01692-X}
}
\usage{
data("glioma.idh.model")
}
\description{
A RF model able to classify DNA methylation samples in to IDH groups
}
\keyword{datasets}
