% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{data}
\name{linkedOmics.data}
\alias{linkedOmics.data}
\title{linkedOmics table}
\format{A table}
\source{
Parsed from 
\url{http://linkedomics.org/login.php#dataSource}
}
\usage{
data("linkedOmics.data")
}
\description{
linkedOmics table with links
}
\keyword{datasets}
