% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{maf.tumor}
\alias{maf.tumor}
\title{TCGA projects with MAF files}
\format{A list with 33 elements}
\usage{
maf.tumor
}
\description{
TCGA projects with MAF files
}
\examples{
\dontrun{
data("maf.tumor")
}
}
\keyword{internal}
