% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{data}
\name{probes2rm}
\alias{probes2rm}
\title{List of probs that should be removed 
from EPIC array due to different versions of the platform}
\format{A list with 977 elements}
\usage{
probes2rm
}
\description{
List of probs that should be removed 
from EPIC array due to different versions of the platform
}
\examples{
\dontrun{
data("probes2rm")
}
}
\keyword{internal}
