% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{package}
\name{TCGAbiolinksGUI.data}
\alias{TCGAbiolinksGUI.data}
\title{Auxiliary data for TCGAbiolinksGUI package.}
\description{
\strong{Package:} TCGAbiolinksGUI.data provide the necessary data for TCGAbiolinksGUI glioma classifier menu.
 It includes the following objects: 
 
\describe{
 \item{\strong{\link[TCGAbiolinksGUI.data]{glioma.gcimp.model}}}{A train model for GCIMP DNA methylation signatures.}
 \item{\strong{\link[TCGAbiolinksGUI.data]{glioma.idhwt.model}}}{A train model for IDHwt DNA methylation signatures}
 \item{\strong{\link[TCGAbiolinksGUI.data]{glioma.idhmut.model}}}{A train model for IDHmut DNA methylation signatures}
 \item{\strong{\link[TCGAbiolinksGUI.data]{glioma.idh.model}}}{A train model for IDH DNA methylation signatures}
 \item{\strong{\link[TCGAbiolinksGUI.data]{probes2rm}}}{List of probes that should be removed from EPIC array due to different versions of the platform.\cr Source: https://support.illumina.com/downloads/infinium-methylationepic-v1-0-product-files.html }
 \item{\strong{\link[TCGAbiolinksGUI.data]{maf.tumor}}}{TCGA projects with open MAF files retrieved from  the NCI's Genomic Data Commons (GDC). \cr Source: https://gdc-docs.nci.nih.gov/Data/Release_Notes/Manifests/GDC_open_MAFs_manifest.txt}
 \item{\strong{\link[TCGAbiolinksGUI.data]{GDCdisease}}}{The NCI's Genomic Data Commons (GDC) projects list}
}
}
\keyword{utilities}
