% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{package}
\name{TCGAbiolinksGUI.data}
\alias{TCGAbiolinksGUI.data}
\alias{TCGAbiolinksGUI.data-package}
\title{Data for TCGAbiolinksGUI package.}
\description{
TCGAbiolinksGUI.data provide the necessary data for TCGAbiolinksGUI glioma classifier menu.
}
\examples{
# Please see the datasets.
}
\seealso{
\code{\link[TCGAbiolinksGUI.data]{glioma.idh.model}}, \code{\link[TCGAbiolinksGUI.data]{glioma.idhwt.model}}, 
\code{\link[TCGAbiolinksGUI.data]{glioma.idhmut.model}}, \code{\link[TCGAbiolinksGUI.data]{glioma.gcimp.model}},
\code{\link[TCGAbiolinksGUI.data]{probes2rm}}, 
\code{\link[TCGAbiolinksGUI.data]{maf.tumor}}, 
\code{\link[TCGAbiolinksGUI.data]{GDCdisease}}
}
\keyword{utilities}
