% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{data}
\name{GDCdisease}
\alias{GDCdisease}
\title{GDC projects}
\format{A named list with 39 projects}
\source{
Retrivied from GDC API
}
\usage{
GDCdisease
}
\description{
Contains all GDC projects with open data
}
\examples{
data("GDCdisease")
}
\keyword{internal}
