% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{data}
\name{glioma.gcimp.model}
\alias{glioma.gcimp.model}
\title{A train model for gcimp DNA methylation signatures}
\format{A train object}
\usage{
glioma.gcimp.model
}
\description{
A train model for gcimp DNA methylation signatures
}
\examples{
\dontrun{
data("glioma.gcimp.model")
}
}
\keyword{internal}
