% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{data}
\name{glioma.idh.model}
\alias{glioma.idh.model}
\title{A train model for IDH DNA methylation signatures}
\format{A train object}
\usage{
glioma.idh.model
}
\description{
A train model for IDH DNA methylation signatures
}
\examples{
\dontrun{
data("glioma.idh.model")
}
}
\keyword{internal}
