% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{glioma.idhmut.model}
\alias{glioma.idhmut.model}
\title{A train model for IDHmut DNA methylation signatures}
\format{A train object}
\usage{
glioma.idhmut.model
}
\description{
A train model for IDHmut DNA methylation signatures
}
\examples{
\dontrun{
data("glioma.idhmut.model")
}
}
\keyword{internal}
