% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{data}
\name{glioma.idhwt.model}
\alias{glioma.idhwt.model}
\title{A train model for IDHwt DNA methylation signatures}
\format{A train object}
\usage{
glioma.idhwt.model
}
\description{
A train model for IDHwt DNA methylation signatures
}
\examples{
\dontrun{
data("glioma.idhwt.model")
}
}
\keyword{internal}
