% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{data}
\name{glioma.idhwt.model}
\alias{glioma.idhwt.model}
\title{IDHwt RF model}
\format{A random forest model with 430 samples and 843 predictors classifying into 3 classes}
\source{
RF model created from DNA methylation signatures retrieved 
from \url{www.cell.com/cell/abstract/S0092-8674(15)01692-X}
}
\usage{
glioma.idhwt.model
}
\description{
A RF model able to classify DNA methylation samples in to IDHwt groups
}
\examples{
data("glioma.idhwt.model")
}
\keyword{internal}
