% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{data}
\name{maf.tumor}
\alias{maf.tumor}
\title{GDC open MAF files}
\format{A named list with 33 tumors}
\source{
\url{https://gdc-docs.nci.nih.gov/Data/Release_Notes/Manifests/GDC_open_MAFs_manifest.txt}
}
\usage{
maf.tumor
}
\description{
Contains the list of GDC project with open MAF files available
}
\examples{
data("maf.tumor")
}
\keyword{internal}
