% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{data}
\name{probes2rm}
\alias{probes2rm}
\title{EPIC probes removed from newer versions}
\format{
A list with 977 probes
}
\source{
\url{https://support.illumina.com/downloads/infinium-methylationepic-v1-0-product-files.html}
}
\usage{
data("probes2rm")
}
\description{
EPIC probes removed from newer versions that should not be used in the analysis
}
\keyword{datasets}
