% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiolinks.data.R
\docType{package}
\name{TCGAbiolinksGUI.data}
\alias{TCGAbiolinksGUI.data}
\alias{TCGAbiolinksGUI.data-package}
\alias{glioma.gcimp.model}
\alias{glioma.idhmut.model}
\alias{glioma.idhwt.model}
\alias{glioma.idh.model}
\alias{GDCdisease}
\alias{maf.tumor}
\alias{probes2rm}
\title{Auxiliary data for TCGAbiolinksGUI package.}
\format{A train object}
\usage{
glioma.gcimp.model

glioma.idhmut.model

glioma.idhwt.model

glioma.idh.model

GDCdisease

maf.tumor

probes2rm
}
\description{
TCGAbiolinksGUI.data provide the necessary data for TCGAbiolinksGUI glioma classifier menu.

A train model for gcimp DNA methylation signatures

A train model for IDHmut DNA methylation signatures

A train model for IDHwt DNA methylation signatures

A train model for IDH DNA methylation signatures

GDC projects list

TCGA projects with MAF files

List of probs that should be removed 
from EPIC array due to different versions of the platform
Source: https://support.illumina.com/downloads/infinium-methylationepic-v1-0-product-files.html
}
\examples{
# Please see the datasets.
data("glioma.gcimp.model")
data("glioma.idhmut.model")
data("glioma.idhwt.model")
data("glioma.idh.model")
data("GDCdisease")
data("maf.tumor")
data("probes2rm")
}
\seealso{
\code{\link[TCGAbiolinksGUI.data]{glioma.idh.model}}, \code{\link[TCGAbiolinksGUI.data]{glioma.idhwt.model}}, 
\code{\link[TCGAbiolinksGUI.data]{glioma.idhmut.model}}, \code{\link[TCGAbiolinksGUI.data]{glioma.gcimp.model}},
\code{\link[TCGAbiolinksGUI.data]{probes2rm}}, 
\code{\link[TCGAbiolinksGUI.data]{maf.tumor}}, 
\code{\link[TCGAbiolinksGUI.data]{GDCdisease}}
}
\keyword{internal}
\keyword{utilities}
